+++ project/src/main/java/com/onlybuns/OnlyBuns/controller/api/RestController_Post.java
@@ -22,7 +22,7 @@ public class RestController_Post {
     public ResponseEntity<List<DTO_Get_Post>> get_api_posts(@RequestParam(value = "sort", required = false) String sort) { return servicePost.get_api_posts(sort); }
 
     @GetMapping("/api/posts/{id}")
-    public ResponseEntity<DTO_Get_Post> get_api_posts_id(@PathVariable(name = "id") Integer id) {
+    public ResponseEntity<DTO_Get_Post> get_api_posts_id(@PathVariable(name = "id") Long id) {
         return servicePost.get_api_posts_id(id);
     }
     @PostMapping("/api/createpost")
@@ -35,12 +35,12 @@ public class RestController_Post {
     }
 
     @GetMapping("/api/posts/{id}/replies")
-    public ResponseEntity<List<DTO_Get_Post>> get_api_posts_id_replies(@PathVariable(name = "id") Integer id) {
+    public ResponseEntity<List<DTO_Get_Post>> get_api_posts_id_replies(@PathVariable(name = "id") Long id) {
         return servicePost.get_api_posts_id_replies(id);
     }
     @PostMapping("/api/posts/{id}/replies")
     public ResponseEntity<String> post_api_posts_id_replies(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             @RequestBody Post reply,
             HttpSession session) {
 
@@ -49,12 +49,12 @@ public class RestController_Post {
 
 
     @PostMapping("/api/posts/{id}/like")
-    public ResponseEntity<String> post_api_posts_id_like(@PathVariable(name = "id") Integer id, HttpSession session) {
+    public ResponseEntity<String> post_api_posts_id_like(@PathVariable(name = "id") Long id, HttpSession session) {
         return servicePost.post_api_posts_id_like(id, session);
     }
 
     @PostMapping("/api/posts/{id}/likes")
-    public ResponseEntity<List<DTO_Get_Like>> post_api_posts_id_likes(@PathVariable(name = "id") Integer id) {
+    public ResponseEntity<List<DTO_Get_Like>> post_api_posts_id_likes(@PathVariable(name = "id") Long id) {
         return servicePost.post_api_posts_id_likes(id);
     }
 
@@ -63,7 +63,7 @@ public class RestController_Post {
     // TODO: UPDATE POST
     @PutMapping("/api/posts/{id}")
     public ResponseEntity<String> put_api_posts_id(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             DTO_Put_Post dto_put_post,
             HttpSession session) {
 
@@ -73,7 +73,7 @@ public class RestController_Post {
     // TODO: DELETE POST
     @DeleteMapping("/api/posts/{id}")
     public ResponseEntity<String> delete_api_posts_id(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             HttpSession session) {
 
         return servicePost.delete_api_posts_id(id, session);
